renaming fpa to ffs
