+++ exe/fil-tools/cmd/fpa_add.go
@@ -13,15 +13,15 @@ import (
 )
 
 func init() {
-	fpaAddCmd.Flags().StringP("token", "t", "", "FPA access token")
+	ffsAddCmd.Flags().StringP("token", "t", "", "FFS access token")
 
-	fpaCmd.AddCommand(fpaAddCmd)
+	ffsCmd.AddCommand(ffsAddCmd)
 }
 
-var fpaAddCmd = &cobra.Command{
+var ffsAddCmd = &cobra.Command{
 	Use:   "add [cid|path]",
-	Short: "Add data to FPA via cid or file path",
-	Long:  `Add data to FPA via a cid already in IPFS or local file path`,
+	Short: "Add data to FFS via cid or file path",
+	Long:  `Add data to FFS via a cid already in IPFS or local file path`,
 	PreRun: func(cmd *cobra.Command, args []string) {
 		err := viper.BindPFlags(cmd.Flags())
 		checkErr(err)
@@ -39,7 +39,7 @@ var fpaAddCmd = &cobra.Command{
 		if token == "" {
 			Fatal(errors.New("add requires token"))
 		}
-		ctx = context.WithValue(ctx, authKey("fpatoken"), token)
+		ctx = context.WithValue(ctx, authKey("ffstoken"), token)
 
 		var c *cid.Cid
 		var f *os.File
@@ -56,19 +56,19 @@ var fpaAddCmd = &cobra.Command{
 		}
 
 		if c != nil {
-			s := spin.New("%s Adding specified cid to FPA...")
+			s := spin.New("%s Adding specified cid to FFS...")
 			s.Start()
-			err = fcClient.Fpa.AddCid(ctx, *c)
+			err = fcClient.Ffs.AddCid(ctx, *c)
 			s.Stop()
 			checkErr(err)
-			Success("Added data for cid %s to FPA", c.String())
+			Success("Added data for cid %s to FFS", c.String())
 		} else if f != nil {
-			s := spin.New("%s Adding specified file to FPA...")
+			s := spin.New("%s Adding specified file to FFS...")
 			s.Start()
-			cid, err := fcClient.Fpa.AddFile(ctx, f)
+			cid, err := fcClient.Ffs.AddFile(ctx, f)
 			s.Stop()
 			checkErr(err)
-			Success("Added file to FPA with resulting cid: %s", cid.String())
+			Success("Added file to FFS with resulting cid: %s", cid.String())
 		} else {
 			Fatal(errors.New("you must provide a valid cid or file path"))
 		}
