+++ ffs/pb/fps.proto
@@ -1,0 +1,97 @@
+syntax = "proto3";
+package filecoin.ffs.pb;
+
+option java_multiple_files = true;
+option java_package = "io.textile.filecoin.ffs.pb";
+option java_outer_classname = "Filecoinffs";
+option objc_class_prefix = "TTE";
+
+message AddCidRequest {
+    string cid = 1;
+}
+
+message AddCidReply {
+}
+
+message AddFileRequest {
+   bytes chunk = 1;
+}
+
+message AddFileReply {
+   string cid = 1;
+}
+
+message GetRequest {
+    string cid = 1;
+}
+
+message GetReply {
+    bytes chunk = 1;
+}
+
+message CreateRequest {
+
+}
+
+message CreateReply {
+    string id = 1;
+    string address = 2;
+}
+
+message ShowRequest {
+   string cid = 1;
+}
+
+message InfoRequest {
+}
+
+message InfoReply {
+   string id = 1;
+   WalletInfo wallet = 2;
+   repeated string pins = 3;
+}
+
+message WalletInfo {
+   string address = 1;
+   uint64 balance = 2;
+}
+
+message ShowReply {
+   string cid = 1;
+   int64 created = 2;
+   HotInfo hot = 3;
+   ColdInfo cold = 4;
+   
+   message HotInfo {
+      int64 size = 1;
+      IpfsHotInfo ipfs = 2;
+   }
+
+   message IpfsHotInfo {
+      int64 created = 1;
+   }
+
+   message ColdInfo {
+      FilInfo filecoin = 1;
+   }
+
+   message FilInfo {
+      string payloadCid = 1;
+      uint64 duration = 2;
+      repeated FilStorage proposals = 3;
+   }
+
+   message FilStorage {
+      string ProposalCid = 1;
+      bool Failed = 2;
+   }
+}
+
+service API {
+    rpc AddCid(AddCidRequest) returns (AddCidReply) {}
+    rpc AddFile(stream AddFileRequest) returns (AddFileReply) {}
+    rpc Get(GetRequest) returns (stream GetReply) {}
+    rpc Create(CreateRequest) returns (CreateReply) {}
+    rpc Info(InfoRequest) returns (InfoReply) {}
+    rpc Show(ShowRequest) returns (ShowReply) {}
+}
